+++ src/stores/goal.ts
@@ -1,27 +1,37 @@
+import { computed, ref } from 'vue';
 import { defineStore } from 'pinia';
+import { useEventBus } from '@vueuse/core';
 
 import { cmpGoalByCompletion } from 'src/lib/goal.ts';
-import { getGoals, GoalWithAchievement } from 'src/lib/api/goal.ts';
+import { getGoals, Goal, GoalWithAchievement } from 'src/lib/api/goal.ts';
 
-export const useGoalStore = defineStore('goal', {
-  state: () : { goals: GoalWithAchievement[] | null; } => {
-    return { goals: null };
-  },
-  getters: {
-    starredGoals: state => {
-      if(state.goals === null) {
-        return [];
-      }
+export const useGoalStore = defineStore('goal', () => {
+  const goals = ref<GoalWithAchievement[] | null>(null);
 
-      return state.goals.filter(goal => goal.starred);
-    },
-  },
-  actions: {
-    async populate(force = false) {
-      if(force || this.goals === null) {
-        const goals = await getGoals();
-        this.goals = goals.sort(cmpGoalByCompletion);
-      }
+  function $reset() {
+    goals.value = null;
+  }
+
+  const allGoals = computed(() => goals.value ?? []);
+  const starredGoals = computed(() => allGoals.value.filter(goal => goal.starred));
+
+  async function populate(force = false) {
+    if(force || goals.value === null) {
+      const fetchedWorks = await getGoals();
+      goals.value = fetchedWorks.sort(cmpGoalByCompletion);
     }
-  },
+  }
+
+  useEventBus<{ goal: Goal }>('goal:create').on(() => populate(true));
+  useEventBus<{ goal: Goal }>('goal:edit').on(() => populate(true));
+  useEventBus<{ goal: Goal }>('goal:star').on(() => populate(true));
+  useEventBus<{ goal: Goal }>('goal:delete').on(() => populate(true));
+
+  return {
+    goals,
+    allGoals,
+    starredGoals,
+    populate,
+    $reset,
+  };
 });